feat(ui): refactor metric history modals to use MetricHistoryView component
